Tests/LibWeb/Text/expected/HTML/HTMLMediaElement-screen-wake-lock.txt

initial: screenWakeLockActive=false
loaded metadata: screenWakeLockActive=false
playing: screenWakeLockActive=true
paused: screenWakeLockActive=false
playing again: screenWakeLockActive=true
video track disabled: screenWakeLockActive=false
video track re-enabled: screenWakeLockActive=true
src cleared and loaded: screenWakeLockActive=false
